Refactor video session resources to a separate module
