+++ vk-video/src/vulkan_decoder/wrappers.rs
@@ -5,6 +5,7 @@ use ash::Entry;
 mod command;
 mod debug;
 mod mem;
+mod parameter_sets;
 mod sync;
 mod video;
 mod vk_extensions;
@@ -12,6 +13,7 @@ mod vk_extensions;
 pub(crate) use command::*;
 pub(crate) use debug::*;
 pub(crate) use mem::*;
+pub(crate) use parameter_sets::*;
 pub(crate) use sync::*;
 pub(crate) use video::*;
 pub(crate) use vk_extensions::*;
